feature: implement system notifications, add user nickname management, and enhance translation functionality

+++ composeApp/src/commonMain/kotlin/io/rudione/chatone/data/remote/dto/SevenTvCosmeticsDto.kt
@@ -134,7 +134,7 @@ data class SevenTvCosmeticsResponse(
 data class SevenTvEventMessage(
     val op: Int,
     val d: SevenTvEventData? = null,
-    val t: Int? = null
+    val t: Long? = null
 )
 
 @Serializable
